Bug 1447124 - Use int64_t for SaferMultDiv. r=gerald
This prevent potential division by zero should the cast on the argument cause an overflow.
We still limit the mul and div arguments to INT64_MAX.

MozReview-Commit-ID: gHkv6m4zq0

UltraBlame original commit: aff1318be50a9ad35a9dba3198f599dccf725dc3

+++ dom/media/VideoUtils.h
@@ -126,7 +126,7 @@ CheckedInt64 FramesToUsecs(int64_t aFrames, uint32_t aRate);
 media::TimeUnit FramesToTimeUnit(int64_t aFrames, uint32_t aRate);
 
 
-CheckedInt64 SaferMultDiv(int64_t aValue, uint32_t aMul, uint32_t aDiv);
+CheckedInt64 SaferMultDiv(int64_t aValue, uint64_t aMul, uint64_t aDiv);
